+++ core/framework/src/test/java/org/phoebus/framework/nls/NLSMessagesTest.java
@@ -7,15 +7,22 @@
  ******************************************************************************/
 package org.phoebus.framework.nls;
 
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FilenameFilter;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Enumeration;
+import java.util.List;
+import java.util.Locale;
+import java.util.Properties;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
-import java.util.Locale;
-
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.MatcherAssert.assertThat;
 
 
 /** JUnit test of the {@link NLS} message initialization
@@ -31,6 +38,7 @@ public class NLSMessagesTest
     public static String MissingMessage;
     public static String AppVersion;
     private static Locale original;
+    private static final String MESSAGE = "messages";
 
     // They are initialized from a "messages*.properties" file in the same package,
     // typically using 'static' code like this:
@@ -95,7 +103,8 @@ public class NLSMessagesTest
     @Test
     public void testUpdateMessages() {
         System.out.println("**compare messages.properties files***");
-        List<String> difference = NLS.checkMessageFilesDifferences(NLSMessagesTest.class);
+        URL resource = NLSMessagesTest.class.getResource(MESSAGE + ".properties");
+        List<String> difference = NLSMessagesTest.checkMessageFilesDifferences(resource.getFile());
         System.out.println("**There is " + (difference != null ? difference.size() : 0) + " difference(s) found**");
         boolean appversionFound = false;
         if (difference != null && !difference.isEmpty()) {
@@ -111,7 +120,7 @@ public class NLSMessagesTest
 
     @Test
     public void checkAllMessagesResources() {
-        List<String> differences = NLS.checkAllMessageFilesDifferences();
+        List<String> differences = NLSMessagesTest.checkAllMessageFilesDifferences();
         if (differences != null && !differences.isEmpty()) {
             System.out.println("WARNING ! there is " + differences.size() + " difference(s) detected in Messages_{LOCAL}.properties files");
         }
@@ -122,4 +131,170 @@ public class NLSMessagesTest
     public static void restoreLocale() {
         Locale.setDefault(original);
     }
+   
+    
+    /**
+     * Use for unit test only
+     * Check if all the existing messages_{LOCALE}.properties are synchronized on default messages.propertiesresource in the project
+     * 
+     * @return the list of difference between the default resources , null or empty if it is synchronized
+     */
+    public static List<String> checkAllMessageFilesDifferences(){
+        List<String> differences = new ArrayList<>();
+        URL resource = NLSMessagesTest.class.getResource(MESSAGE + ".properties");
+        if (resource != null) {
+            String filePath = resource.getFile();
+            System.out.println(filePath);
+            String[] split = filePath.split("/core/framework/");
+            //First part is the parent folder
+            String parentFolder = split != null && split.length > 0 ? split[0] : null;
+            System.out.println("parentFolder=" + parentFolder);
+            File parentFile = new File(parentFolder);
+            List<File> fileList = listMessagesFiles(parentFile);
+           
+            for (File file : fileList) {
+                List<String> diff = NLSMessagesTest.checkMessageFilesDifferences(file.getAbsolutePath());
+                if (diff != null && !diff.isEmpty()) {
+                    differences.addAll(diff);
+                }
+            }
+
+            if(differences.isEmpty()) {
+                System.out.println("All the "+ MESSAGE+ "_{LOCALE}.properties files are syncronized ");
+            }
+            else {
+                System.out.println("**There is " + differences.size() + " difference(s) found**");
+                for (String dif : differences) {
+                    System.out.println(dif);
+                }
+            }
+        }
+        return differences;
+    }
+    
+    /**
+     * Use for unit test only
+     * Check if the existing messages_{LOCALE}.properties are synchronized on default messages.propertiesresource
+     * 
+     * @param clazz Class relative to which message resources are located
+     * @return the list of difference between the default ressources , null or empty if it is synchronized
+     */
+    private static List<String> checkMessageFilesDifferences(String resourceFile) {
+        List<String> differences = new ArrayList<>();
+        if (resourceFile != null) {
+            try {
+                File defaultFile = new File(resourceFile);
+                Properties defaultBundle = new Properties();
+                defaultBundle.load(new FileInputStream(defaultFile));
+                File parent = defaultFile.getParentFile();
+                FilenameFilter fileNameFilter = new FilenameFilter() {
+
+                    @Override
+                    public boolean accept(File dir, String name) {
+                        return name.startsWith(MESSAGE) && name.endsWith(".properties");
+                    }
+                };
+
+                File[] listFiles = parent.listFiles(fileNameFilter);
+                if (listFiles != null && listFiles.length > 0) {
+                    // System.out.println("Number of languages found =" + listFiles.length);
+                    String fileName = null;
+                    String countryCode = null;
+                    String countryName = null;
+                    Properties compareBundle = null;
+                    Object key = null;
+                    String value = null;
+                    Locale locale = null;
+                    Enumeration<Object> compareKeys = null;
+                    Enumeration<Object> defaultKeys = null;
+                    for (File tmpFile : listFiles) {
+                        fileName = tmpFile.getName();
+                        // Do not compare to itself
+                        if (!fileName.equalsIgnoreCase(defaultFile.getName())) {
+                            // Extract the country code
+                            countryCode = fileName.replaceFirst(MESSAGE + "_", "");
+                            countryCode = countryCode.replace(".properties", "");
+                            locale = getLocaleFromCountryCode(countryCode);
+                            if (locale != null) {
+                                countryName = locale.getDisplayCountry();
+                                compareBundle = new Properties();
+                                compareBundle.load(new FileInputStream(tmpFile));
+                                // Check if the key exist in the LOCAL file
+                                // System.out.println("Compare " + tmpFile.getName() + " to " + defaultFile.getName());
+                                defaultKeys = defaultBundle.keys();
+                                while (defaultKeys.hasMoreElements()) {
+                                    key = defaultKeys.nextElement();
+                                    if (!compareBundle.containsKey(key)) {
+                                        // Ignore env variables eg ${revision}
+                                        value = String.valueOf(defaultBundle.get(key));
+                                        if (!(value.startsWith("${") && value.endsWith("}"))) {
+                                            differences.add("Missing " + key + " in " + countryName + " resource "
+                                                    + tmpFile.getAbsolutePath());
+                                        }
+                                    }
+                                }
+
+                                // Check if there are some key to remove in LOCAL file
+                                compareKeys = compareBundle.keys();
+                                while (compareKeys.hasMoreElements()) {
+                                    key = compareKeys.nextElement();
+                                    if (!defaultBundle.containsKey(key)) {
+                                        differences.add("Remove " + key + " in " + countryName + " resource "
+                                                + tmpFile.getAbsolutePath());
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return differences;
+    }
+    
+    private static List<File> listMessagesFiles(File folder) {
+        String filename = MESSAGE + ".properties";
+        List<File> fileList = new ArrayList<>();
+        //Ignore target folder from build
+        if(folder != null && folder.isDirectory() 
+                && !folder.getAbsolutePath().contains("\\target\\")
+                && !folder.getAbsolutePath().contains("\\test\\")) {
+            File[] listFiles = folder.listFiles();
+            for(File file : listFiles) {
+                if(file.isDirectory()) {
+                    List<File> list = listMessagesFiles(file);
+                    fileList.addAll(list);
+                }
+                else if (file.getName().equals(filename)){
+                    fileList.add(file);
+                }
+            }
+        }
+        return fileList;
+    }
+     
+    
+    /**
+     * To get Locale from a countryCode
+     * 
+     * @param countryCode (fr , en ..."
+     * @return Locale
+     */
+    private static Locale getLocaleFromCountryCode(String countryCode) {
+        Locale localFound = null;
+        if (countryCode != null && !countryCode.isEmpty()) {
+            Locale[] availableLocales = Locale.getAvailableLocales();
+            for (Locale locale : availableLocales) {
+                if (locale.getCountry().toLowerCase().equals(countryCode.toLowerCase())) {
+                    localFound = locale;
+                    break;
+                }
+            }
+        }
+        return localFound;
+    }
+
 }
